% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalEqualRiskContribution.R
\name{hierarchicalEqualRiskContribution}
\alias{hierarchicalEqualRiskContribution}
\title{Design of hierarchical equal risk contribution portfolios}
\usage{
hierarchicalEqualRiskContribution(
  asset_prices = NULL,
  asset_returns = NULL,
  Sigma = NULL,
  risk_measure = c("variance", "standard-deviation", "equal-weighting", "CVaR", "CDaR"),
  linkage = "ward.D2",
  num_clusters = NULL
)
}
\arguments{
\item{asset_prices}{An XTS object of the asset prices.}

\item{asset_returns}{An XTS object of the asset returns.}

\item{Sigma}{Covariance matrix of returns. If none is provided, the
covariance matrix will be computed from the returns.}

\item{risk_measure}{String indicating the desired risk measure for assigning
portfolio weights. Must be one of c('variance', 'standard-deviation',
'equal-weighting', 'CVaR', 'CDaR')}

\item{linkage}{String indicating the desired linkage function. Must be one
of c("single", "complete","average" ,"ward.D", "ward.D2" )}

\item{num_clusters}{Integer value representing the optimal number of clusters.
If no value is given, the optimal number of clusters will be computed
automatically.}
}
\description{
This function designs hierarchical equal risk contribution
portfolios based on the method proposed by Raffinot (2018).
}
\details{
This portfolio allocation method makes use of hierarchical clustering
to assign portfolio weights.
}
