% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalRiskParity.R
\name{hierarchicalRiskParity}
\alias{hierarchicalRiskParity}
\title{Design of hierarchical risk parity portfolios}
\usage{
hierarchicalRiskParity(
  asset_prices = NULL,
  asset_returns = NULL,
  Sigma = NULL,
  linkage = "single"
)
}
\arguments{
\item{asset_prices}{An XTS object of the asset prices.}

\item{asset_returns}{An XTS object of the asset returns.}

\item{Sigma}{Covariance matrix of returns. If none is provided, the
covariance matrix will be computed from the returns.}

\item{linkage}{String indicating the desired linkage function. Must be one
of c("single", "complete","average" ,"ward.D", "ward.D2" )}
}
\description{
This function designs a hierarchical risk parity portfolio
based on Lopez de Prado's paper "Building Diversified Portfolios that
Outperform Out-of-Sample" (2015).
}
\details{
This portfolio allocation method is a heuristic method which makes
use of hierarchical clustering to seriate the correlation matrix of asset
returns, then recursively allocates the portfolio weights via naive risk
parity and "recursive bisection".
}
