% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHRP.R
\name{NHRP}
\alias{NHRP}
\title{Nodal Hierarchical Risk Parity}
\usage{
NHRP(
  asset_prices = NULL,
  asset_returns = NULL,
  Sigma = NULL,
  risk_measure = c("variance", "standard-deviation", "CVaR", "CDaR"),
  method = "ward.D2",
  w_min = NULL,
  w_max = NULL,
  lam = 1
)
}
\arguments{
\item{asset_prices}{An XTS object of the asset prices.}

\item{asset_returns}{An XTS object of the asset returns.}

\item{Sigma}{Covariance matrix of returns. If none is provided, the
covariance matrix will be computed from the returns.}

\item{risk_measure}{String indicating the desired risk measure for assigning
portfolio weights. Must be one of c('variance', 'standard-deviation',
'CVaR', 'CDaR')}

\item{method}{String indicating the desired hierarchical clustering method.
Must be one of c("complete", "single", "average" ,"ward.D", "ward.D2",
"divisive"). If method="divisive", divisive clustering (or the DIANA
algorithm)is used, otherwise agglomerative clustering is used with
method referring to the desired linkage function. The default is
"complete".}

\item{w_min}{Scalar or vector with values between \link{0,1} to control the
minimum value of weights.}

\item{w_max}{Scalar or vector with values between \link{0,1} to control the
maximum value of weights.}

\item{lam}{Non-negative tuning parameter to control the concentration into
different clusters.}
}
\description{
A modified version of Lopez de Prado's (2015) Hierarchical Risk
Parity (HRP) method which incorporates the hierarchical structure of the minimum
spanning tree obtained from hierarchical clustering.
}
\details{
This portfolio allocation method is an extension of the
Hierarchical Risk Parity (HRP) method which learns the minimum spanning tree
from historical asset returns, then recursively allocates the portfolio
weights via naive risk parity and "recursive bisection" according to the
tree structure.
}
